Improve shader recompilation logic

+++ init_fafbeta.lua
@@ -411,7 +411,9 @@ end
 -- Clears out the shader cache as it takes a release to reset the shaders
 local shaderCache = SHGetFolderPath('LOCAL_APPDATA') .. 'Gas Powered Games/Supreme Commander Forged Alliance/cache'
 for k, file in IoDir(shaderCache .. '/*') do
-    os.remove(shaderCache .. '/' .. file)
+    if file != '.' and file != '..' then 
+        os.remove(shaderCache .. '/' .. file)
+    end
 end
 
 -- typical FAF packages
